+++ test/notification/apsProperties.js
@@ -261,6 +261,7 @@ describe('Notification', function () {
         });
       });
     });
+
     describe('titleLocKey', function () {
       it('sets the aps.alert.title-loc-key property', function () {
         note.titleLocKey = 'Warning';
@@ -349,6 +350,108 @@ describe('Notification', function () {
         it('sets the aps.alert.title-loc-args property', function () {
           expect(compiledOutput())
             .to.have.nested.deep.property('aps.alert.title-loc-args')
+            .that.deep.equals(['Hi there']);
+        });
+      });
+
+      describe('setTitleLocArgs', function () {
+        it('is chainable', function () {
+          expect(note.setTitleLocArgs(['iPhone 6s'])).to.equal(note);
+          expect(compiledOutput())
+            .to.have.nested.deep.property('aps.alert.title-loc-args')
+            .that.deep.equals(['iPhone 6s']);
+        });
+      });
+    });
+
+    describe('subtitleLocKey', function () {
+      it('sets the aps.alert.subtitle-loc-key property', function () {
+        note.subtitleLocKey = 'Warning';
+        expect(compiledOutput()).to.have.nested.deep.property('aps.alert.subtitle-loc-key', 'Warning');
+      });
+
+      context('alert is already an object', function () {
+        beforeEach(function () {
+          note.alert = { body: 'Test', 'launch-image': 'test.png' };
+          note.subtitleLocKey = 'Warning';
+        });
+
+        it('contains all expected properties', function () {
+          expect(compiledOutput()).to.have.nested.deep.property('aps.alert').that.deep.equals({
+            body: 'Test',
+            'launch-image': 'test.png',
+            'subtitle-loc-key': 'Warning',
+          });
+        });
+      });
+
+      context('alert is already a string', function () {
+        beforeEach(function () {
+          note.alert = 'Hello, world';
+          note.subtitleLocKey = 'Warning';
+        });
+
+        it('retains the alert body correctly', function () {
+          expect(compiledOutput()).to.have.nested.deep.property('aps.alert.body', 'Hello, world');
+        });
+
+        it('sets the aps.alert.subtitle-loc-key property', function () {
+          expect(compiledOutput()).to.have.nested.deep.property(
+            'aps.alert.subtitle-loc-key',
+            'Warning'
+          );
+        });
+      });
+
+      describe('setAlert', function () {
+        it('is chainable', function () {
+          expect(note.setSubtitleLocKey('greeting')).to.equal(note);
+          expect(compiledOutput()).to.have.nested.deep.property(
+            'aps.alert.subtitle-loc-key',
+            'greeting'
+          );
+        });
+      });
+    });
+
+    describe('subtitleLocArgs', function () {
+      it('sets the aps.alert.subtitle-loc-args property', function () {
+        note.subtitleLocArgs = ['arg1', 'arg2'];
+        expect(compiledOutput())
+          .to.have.nested.deep.property('aps.alert.subtitle-loc-args')
+          .that.deep.equals(['arg1', 'arg2']);
+      });
+
+      context('alert is already an object', function () {
+        beforeEach(function () {
+          note.alert = { body: 'Test', 'launch-image': 'test.png' };
+          note.subtitleLocArgs = ['Hi there'];
+        });
+
+        it('contains all expected properties', function () {
+          expect(compiledOutput())
+            .to.have.nested.deep.property('aps.alert')
+            .that.deep.equals({
+              body: 'Test',
+              'launch-image': 'test.png',
+              'subtitle-loc-args': ['Hi there'],
+            });
+        });
+      });
+
+      context('alert is already a string', function () {
+        beforeEach(function () {
+          note.alert = 'Hello, world';
+          note.subtitleLocArgs = ['Hi there'];
+        });
+
+        it('retains the alert body', function () {
+          expect(compiledOutput()).to.have.nested.deep.property('aps.alert.body', 'Hello, world');
+        });
+
+        it('sets the aps.alert.subtitle-loc-args property', function () {
+          expect(compiledOutput())
+            .to.have.nested.deep.property('aps.alert.subtitle-loc-args')
             .that.deep.equals(['Hi there']);
         });
       });
@@ -832,28 +935,54 @@ describe('Notification', function () {
       });
     });
 
-    describe('events', function () {
+    describe('event', function () {
       it('defaults to undefined', function () {
-        expect(compiledOutput()).to.not.have.nested.property('aps.events');
+        expect(compiledOutput()).to.not.have.nested.property('aps.event');
       });
 
       it('can be set to a string', function () {
-        note.events = 'the-event';
+        note.event = 'the-event';
 
-        expect(compiledOutput()).to.have.nested.property('aps.events', 'the-event');
+        expect(compiledOutput()).to.have.nested.property('aps.event', 'the-event');
       });
 
       it('can be set to undefined', function () {
-        note.events = 'the-event';
-        note.events = undefined;
+        note.event = 'the-event';
+        note.event = undefined;
 
-        expect(compiledOutput()).to.not.have.nested.property('aps.events');
+        expect(compiledOutput()).to.not.have.nested.property('aps.event');
       });
 
-      describe('setEvents', function () {
+      describe('setEvent', function () {
         it('is chainable', function () {
-          expect(note.setEvents('the-event')).to.equal(note);
-          expect(compiledOutput()).to.have.nested.property('aps.events', 'the-event');
+          expect(note.setEvent('the-event')).to.equal(note);
+          expect(compiledOutput()).to.have.nested.property('aps.event', 'the-event');
+        });
+      });
+    });
+
+    describe('dismissal-date', function () {
+      it('defaults to undefined', function () {
+        expect(compiledOutput()).to.not.have.nested.property('aps.dismissal-date');
+      });
+
+      it('can be set to a number', function () {
+        note.dismissalDate = 123456;
+
+        expect(compiledOutput()).to.have.nested.property('aps.dismissal-date', 123456);
+      });
+
+      it('can be set to undefined', function () {
+        note.dismissalDate = 123456;
+        note.dismissalDate = undefined;
+
+        expect(compiledOutput()).to.not.have.nested.property('aps.dismissal-date');
+      });
+
+      describe('setDismissalDate', function () {
+        it('is chainable', function () {
+          expect(note.setDismissalDate(123456)).to.equal(note);
+          expect(compiledOutput()).to.have.nested.property('aps.dismissal-date', 123456);
         });
       });
     });